Example Flowchart
Considering the guessing game example (A computer picks a number between 1 and 100 and a user tries to guess the number) we looked at an example flow of how the game would work.

This code was written to test the concept, and the flowchart was created to display how the code flowed.